'cleaning'
Calls to `n` are now calls to `eta` for readability. `bisect()` changed to `bisectE()` to fix all the little things that have to be changed using the original. `f`, the `bisectE()` output is preallocated, and we don't need `fx`. Calls to `xlim()` are not needed as the new domain `eta` is fixed by the zero-finding method rather than `linspace()` in f. Plot `hold ` calls are moved to before the first plot, so that `hold ` may be simply left `on` after the first plot, wil no need for calls to `clf`.