I.h and II.i
Edits to `res()` and `dydx()` for use in both I.h and II.i, I.h is closer and II.i overshoots into -1.6e-16 on the last timestep.
And there's a ton of warnings from MATLAB on `ode45()` during `res()` concerning the accuracy. If we revert something, it may be this.